feat(RandomAnswer): implement random number answer
