+++ README.md
@@ -3,10 +3,10 @@
 ## 🚀 기능 요구사항
 - [x] 사용자에게 숫자 입력받기
 
-- [ ] (예외) 사용자의 입력이 이상한지 체크
+- [x] (예외) 사용자의 입력이 이상한지 체크
   2. 입력이 3자리인지 체크
   4. 숫자들이 서로다른 숫자인치 체크
-- [ ] 세자리의 서로다른 랜덤넘버 생성하기
+- [x] 세자리의 서로다른 랜덤넘버 생성하기
 - [ ] (로직)입력된 숫자 변수에 저장후 로직에 따른 결과값 출력하기
   1. 틀린경우: 볼, 스트라이크, 낫싱 을 결과값으로 출력
   2. 맞힌경우: "3스트라이크"라는 메세지와 새로 게임할 지 종료할지 선택
