+++ src/main/java/baseball/Application.java
@@ -4,6 +4,9 @@ import java.util.Scanner;
 public class Application {
     public static void main(String[] args) {
         boolean playing = true;
+        RandomAnswer newRandomAnswer = new RandomAnswer();
+        Integer answer = newRandomAnswer.randomAnswer();
+        System.out.println("answer" + answer);
 
         final Scanner scanner = new Scanner(System.in);
         Game newGame = new Game();
