+++ src/main/java/baseball/Game.java
@@ -1,34 +1,28 @@
 package baseball;
 
+import utils.RandomUtils;
+
 import java.util.Scanner;
 
 public class Game {
     public void start(Scanner scanner) {
         Integer number;
 
+
         System.out.print("서로 다른 3자리 숫자를 입력해주세요 : ");
         if (scanner.hasNextInt()) {
             number = scanner.nextInt();
-            if (validate(number)) {
+            Validate myValidate = new Validate();
+            if (myValidate.validate(number)) {
                 System.out.println("입력숫자: " + number);
             }
+            System.out.println("패키지 연결");
+            RandomUtils myUtils = new RandomUtils();
+            System.out.println(myUtils.nextInt(0, 9));
+
         }
         else if (!scanner.hasNextInt()){
             scanner.next();
         }
-    }
-
-    public boolean validate(int number) {
-        int first = number / 100;
-        int second = (number % 100) / 10;
-        int third = number % 10;
-
-        if (number >= 1000 || number < 100) {
-            return false;
-        }
-        else if (first == second || first == third || second == third) {
-            return false;
-        }
-        return true;
     }
 }
