+++ src/main/java/utils/RandomUtils.java
@@ -5,7 +5,7 @@ import java.util.Random;
 public class RandomUtils {
     private static final Random RANDOM = new Random();
 
-    private RandomUtils() {
+    public RandomUtils() {
     }
 
     public static int nextInt(final int startInclusive, final int endInclusive) {
